requirements.txt
langchain
langchain-community
groq
python-dotenv
pip 
openai